independent training works
Former-commit-id: ed4a58068e289670ba9a9bedbb5f729982b058f4

+++ rllab/algo/ppo.py
@@ -130,7 +130,7 @@ class PPO(BatchPolopt):
             "observations", "advantages", "pdists", "actions"
         ))
 
-        cur_params = policy.get_param_values(trainable=True)
+        cur_params = policy.get_param_values(trainable=True).astype(np.float64)
 
         def evaluate_cost(penalty):
             def evaluate(params):
